ai/aiagent/confluence-agent/agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp>=1.2.1",
    "openai>=1.61.0",
    "python-dotenv>=1.0.1",
    "streamlit>=1.42.0",
]
